requirements.txt
altgraph==0.15
asn1crypto==0.23.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
dnspython==1.15.0
ecdsa==0.13
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-SocketIO==2.9.2
future==0.16.0
greenlet==0.4.12
idna==2.5
itsdangerous==0.24
Jinja2==2.10
json-rpc==1.10.8
m3-cdecimal==2.3
macholib==1.9
MarkupSafe==1.0
pefile==2017.11.5
psutil==5.4.2
pycparser==2.18
pycrypto==2.6.1
PyInstaller==3.3.1
pyOpenSSL==17.5.0
python-engineio==2.0.1
python-slugify==1.2.4
python-socketio==1.8.3
PyYAML==3.12
requests==2.18.2
requests-toolbelt==0.8.0
simplekv==0.11.4
six==1.11.0
SQLAlchemy==1.1.15
tabulate==0.8.2
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12.2
docker
ujson